# apps/nextjs/lib/AgentProtocol/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Agent Protocol",
    "description": "Specification of the API protocol for communication with an agent.",
    "version": "v1"
  },
  "servers": [
    { "url": "http://0.0.0.0:8000", "description": "Agent Protocol API" }
  ],
  "paths": {
...
    "/ap/v1/agent/tasks/{task_id}/steps": {
      "get": {
        "operationId": "listAgentTaskSteps",
        "summary": "List all steps for the specified task.",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "ID of the task.",
            "required": true,
            "schema": { "type": "string" },
            "example": "50da533e-3904-4401-8a07-c49adf88b5eb",
            "x-postman-variables": [{ "type": "load", "name": "task_id" }]
          },
          {
            "name": "current_page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1,
              "minimum": 1
            },
            "example": 2
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10,
              "minimum": 1
            },
            "example": 25
          }
        ],
        "responses": {
          "200": {
            "description": "Returned list of agent's steps for the specified task.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "steps": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "description": "Body of the task request.",
                            "type": "object",
                            "properties": {
                              "input": {
                                "description": "Input prompt for the step.",
                                "type": "string",
                                "example": "Write the words you receive to the file 'output.txt'.",
                                "nullable": true
                              },
                              "additional_input": {
                                "description": "Input parameters for the task step. Any value is allowed.",
                                "type": "object",
                                "example": "{\n\"file_to_refactor\": \"models.py\"\n}"
                              }
                            }
                          },
                          {
                            "type": "object",
                            "required": [
                              "step_id",
                              "task_id",
                              "status",
                              "is_last",
                              "artifacts"
                            ],
                            "properties": {
                              "task_id": {
                                "description": "The ID of the task this step belongs to.",
                                "type": "string",
                                "example": "50da533e-3904-4401-8a07-c49adf88b5eb"
                              },
                              "step_id": {
                                "description": "The ID of the task step.",
                                "type": "string",
                                "example": "6bb1801a-fd80-45e8-899a-4dd723cc602e"
                              },
                              "name": {
                                "description": "The name of the task step.",
                                "type": "string",
                                "example": "Write to file",
                                "nullable": true
                              },
                              "status": {
                                "description": "The status of the task step.",
                                "type": "string",
                                "example": "created",
                                "enum": ["created", "running", "completed"]
                              },
                              "output": {
                                "description": "Output of the task step.",
                                "type": "string",
                                "example": "I am going to use the write_to_file command and write Washington to a file called output.txt <write_to_file('output.txt', 'Washington')",
                                "nullable": true
                              },
                              "additional_output": {
                                "description": "Output that the task step has produced. Any value is allowed.",
                                "type": "object",
                                "example": "{\n\"tokens\": 7894,\n\"estimated_cost\": \"0,24$\"\n}",
                                "nullable": true
                              },
                              "artifacts": {
                                "description": "A list of artifacts that the step has produced.",
                                "type": "array",
                                "items": {
                                  "description": "An Artifact either created by or submitted to the agent.",
                                  "type": "object",
                                  "properties": {
                                    "artifact_id": {
                                      "description": "ID of the artifact.",
                                      "type": "string",
                                      "example": "b225e278-8b4c-4f99-a696-8facf19f0e56"
                                    },
                                    "agent_created": {
                                      "description": "Whether the artifact has been created by the agent.",
                                      "type": "boolean",
                                      "example": false
                                    },
                                    "file_name": {
                                      "description": "Filename of the artifact.",
                                      "type": "string",
                                      "example": "main.py"
                                    },
                                    "relative_path": {
                                      "description": "Relative path of the artifact in the agent's workspace.",
                                      "type": "string",
                                      "example": "python/code/",
                                      "nullable": true
                                    }
                                  },
                                  "required": [
                                    "artifact_id",
                                    "agent_created",
                                    "file_name"
                                  ]
                                },
                                "default": []
                              },
                              "is_last": {
                                "description": "Whether this is the last step in the task.",
                                "type": "boolean",
                                "example": true,
                                "default": false
                              }
                            }
                          }
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total_items": {
                          "description": "Total number of items.",
                          "type": "integer",
                          "example": 42
                        },
                        "total_pages": {
                          "description": "Total number of pages.",
                          "type": "integer",
                          "example": 97
                        },
                        "current_page": {
                          "description": "Current_page page number.",
                          "type": "integer",
                          "example": 1
                        },
                        "page_size": {
                          "description": "Number of items per page.",
                          "type": "integer",
                          "example": 25
                        }
                      },
                      "required": [
                        "total_items",
                        "total_pages",
                        "current_page",
                        "page_size"
                      ]
                    }
                  },
                  "required": ["steps", "pagination"]
                }
              }
            }
          },
          "404": {
            "description": "Unable to find entity with a given identifier",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "description": "Message stating the entity was not found",
                      "type": "string",
                      "example": "Unable to find entity with the provided id"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "default": { "description": "Internal Server Error" }
        },
        "tags": ["agent"]
      },
      "post": {
        "operationId": "executeAgentTaskStep",
        "summary": "Execute a step in the specified agent task.",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "ID of the task",
            "required": true,
            "schema": { "type": "string" },
            "example": "50da533e-3904-4401-8a07-c49adf88b5eb",
            "x-postman-variables": [{ "type": "load", "name": "task_id" }]
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "description": "Body of the task request.",
                "type": "object",
                "properties": {
                  "input": {
                    "description": "Input prompt for the step.",
                    "type": "string",
                    "example": "Write the words you receive to the file 'output.txt'.",
                    "nullable": true
                  },
                  "additional_input": {
                    "description": "Input parameters for the task step. Any value is allowed.",
                    "type": "object",
                    "example": "{\n\"file_to_refactor\": \"models.py\"\n}"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Executed step for the agent task.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "description": "Body of the task request.",
                      "type": "object",
                      "properties": {
                        "input": {
                          "description": "Input prompt for the step.",
                          "type": "string",
                          "example": "Write the words you receive to the file 'output.txt'.",
                          "nullable": true
                        },
                        "additional_input": {
                          "description": "Input parameters for the task step. Any value is allowed.",
                          "type": "object",
                          "example": "{\n\"file_to_refactor\": \"models.py\"\n}"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "step_id",
                        "task_id",
                        "status",
                        "is_last",
                        "artifacts"
                      ],
                      "properties": {
                        "task_id": {
                          "description": "The ID of the task this step belongs to.",
                          "type": "string",
                          "example": "50da533e-3904-4401-8a07-c49adf88b5eb"
                        },
                        "step_id": {
                          "description": "The ID of the task step.",
                          "type": "string",
                          "example": "6bb1801a-fd80-45e8-899a-4dd723cc602e"
                        },
                        "name": {
                          "description": "The name of the task step.",
                          "type": "string",
                          "example": "Write to file",
                          "nullable": true
                        },
                        "status": {
                          "description": "The status of the task step.",
                          "type": "string",
                          "example": "created",
                          "enum": ["created", "running", "completed"]
                        },
                        "output": {
                          "description": "Output of the task step.",
                          "type": "string",
                          "example": "I am going to use the write_to_file command and write Washington to a file called output.txt <write_to_file('output.txt', 'Washington')",
                          "nullable": true
                        },
                        "additional_output": {
                          "description": "Output that the task step has produced. Any value is allowed.",
                          "type": "object",
                          "example": "{\n\"tokens\": 7894,\n\"estimated_cost\": \"0,24$\"\n}",
                          "nullable": true
                        },
                        "artifacts": {
                          "description": "A list of artifacts that the step has produced.",
                          "type": "array",
                          "items": {
                            "description": "An Artifact either created by or submitted to the agent.",
                            "type": "object",
                            "properties": {
                              "artifact_id": {
                                "description": "ID of the artifact.",
                                "type": "string",
                                "example": "b225e278-8b4c-4f99-a696-8facf19f0e56"
                              },
                              "agent_created": {
                                "description": "Whether the artifact has been created by the agent.",
                                "type": "boolean",
                                "example": false
                              },
                              "file_name": {
                                "description": "Filename of the artifact.",
                                "type": "string",
                                "example": "main.py"
                              },
                              "relative_path": {
                                "description": "Relative path of the artifact in the agent's workspace.",
                                "type": "string",
                                "example": "python/code/",
                                "nullable": true
                              }
                            },
                            "required": [
                              "artifact_id",
                              "agent_created",
                              "file_name"
                            ]
                          },
                          "default": []
                        },
                        "is_last": {
                          "description": "Whether this is the last step in the task.",
                          "type": "boolean",
                          "example": true,
                          "default": false
                        }
                      }
                    }
                  ]
                }
              }
            },
            "x-postman-variables": [
              {
                "type": "save",
                "name": "step_id",
                "path": ".step_id"
              }
            ]
          },
          "404": {
            "description": "Unable to find entity with a given identifier",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "description": "Message stating the entity was not found",
                      "type": "string",
                      "example": "Unable to find entity with the provided id"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "422": {
            "description": "Unable to process request. Likely due to improperly formatted request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "A generic JSON object without any specific requirements.",
                  "type": "object"
                }
              }
            }
          },
          "default": { "description": "Internal Server Error" }
        },
        "tags": ["agent"]
      }
    },
...
  },
...
}